chore(build): gerar firebase-messaging-sw.js a partir de env vars via plugin Vite
- .firebaserc: aliases prod=bolao-do-bolero e teste=bolao-do-bolero-teste
- vite.config.ts: plugin firebaseMessagingSw serve o SW em dev (middleware) e emite em dist no build, substituindo apiKey/projectId/messagingSenderId/appId de import.meta.env
- remove public/firebase-messaging-sw.js estatico (era hardcoded para prod e impediria separacao entre ambientes)
