+++ vite.config.ts
@@ -1,48 +1,95 @@
 /// <reference types="vitest/config" />
-import { defineConfig } from 'vite'
+import { defineConfig, loadEnv, type Plugin } from 'vite'
 import react from '@vitejs/plugin-react'
 import tailwindcss from '@tailwindcss/vite'
 import { VitePWA } from 'vite-plugin-pwa'
+import { writeFileSync } from 'node:fs'
+import path from 'node:path'
 
-export default defineConfig({
-  plugins: [
-    react(),
-    tailwindcss(),
-    VitePWA({
-      registerType: 'autoUpdate',
-      includeAssets: ['favicon.svg', 'apple-touch-icon.png'],
-      manifest: {
-        name: 'Bolão do Bolero (Duda)',
-        short_name: 'Bolero',
-        description: 'Bolão da Copa do Mundo 2026',
-        lang: 'pt-BR',
-        theme_color: '#1e40af',
-        background_color: '#ffffff',
-        display: 'standalone',
-        orientation: 'portrait',
-        start_url: '/',
-        scope: '/',
-        icons: [
-          { src: '/icon-192.png', sizes: '192x192', type: 'image/png' },
-          { src: '/icon-512.png', sizes: '512x512', type: 'image/png' },
-          { src: '/icon-maskable-192.png', sizes: '192x192', type: 'image/png', purpose: 'maskable' },
-          { src: '/icon-maskable-512.png', sizes: '512x512', type: 'image/png', purpose: 'maskable' },
-        ],
-      },
-      workbox: {
-        globPatterns: ['**/*.{js,css,html,ico,png,svg,woff2}'],
-        navigateFallback: '/index.html',
-        navigateFallbackDenylist: [/^\/api/, /^\/firebase-messaging-sw\.js$/],
-      },
-    }),
-  ],
-  server: {
-    headers: {
-      'Cross-Origin-Opener-Policy': 'same-origin-allow-popups',
+function firebaseMessagingSw(env: Record<string, string>): Plugin {
+  const buildContent = () => `importScripts('https://www.gstatic.com/firebasejs/10.12.0/firebase-app-compat.js')
+importScripts('https://www.gstatic.com/firebasejs/10.12.0/firebase-messaging-compat.js')
+
+firebase.initializeApp({
+  apiKey: '${env.VITE_FIREBASE_API_KEY || ''}',
+  projectId: '${env.VITE_FIREBASE_PROJECT_ID || ''}',
+  messagingSenderId: '${env.VITE_FIREBASE_MESSAGING_SENDER_ID || ''}',
+  appId: '${env.VITE_FIREBASE_APP_ID || ''}',
+})
+
+const messaging = firebase.messaging()
+
+messaging.onBackgroundMessage((payload) => {
+  const { title, body } = payload.notification || {}
+  self.registration.showNotification(title || 'Bolão do Bolero (Duda)', {
+    body,
+    icon: '/icon-192.png',
+  })
+})
+`
+
+  return {
+    name: 'firebase-messaging-sw',
+    configureServer(server) {
+      server.middlewares.use((req, res, next) => {
+        if (req.url === '/firebase-messaging-sw.js') {
+          res.setHeader('Content-Type', 'application/javascript')
+          res.end(buildContent())
+          return
+        }
+        next()
+      })
     },
-  },
-  test: {
-    globals: true,
-    environment: 'jsdom',
-  },
+    writeBundle(options) {
+      const outDir = options.dir || 'dist'
+      writeFileSync(path.join(outDir, 'firebase-messaging-sw.js'), buildContent())
+    },
+  }
+}
+
+export default defineConfig(({ mode }) => {
+  const env = loadEnv(mode, process.cwd(), '')
+  return {
+    plugins: [
+      react(),
+      tailwindcss(),
+      firebaseMessagingSw(env),
+      VitePWA({
+        registerType: 'autoUpdate',
+        includeAssets: ['favicon.svg', 'apple-touch-icon.png'],
+        manifest: {
+          name: 'Bolão do Bolero (Duda)',
+          short_name: 'Bolero',
+          description: 'Bolão da Copa do Mundo 2026',
+          lang: 'pt-BR',
+          theme_color: '#1e40af',
+          background_color: '#ffffff',
+          display: 'standalone',
+          orientation: 'portrait',
+          start_url: '/',
+          scope: '/',
+          icons: [
+            { src: '/icon-192.png', sizes: '192x192', type: 'image/png' },
+            { src: '/icon-512.png', sizes: '512x512', type: 'image/png' },
+            { src: '/icon-maskable-192.png', sizes: '192x192', type: 'image/png', purpose: 'maskable' },
+            { src: '/icon-maskable-512.png', sizes: '512x512', type: 'image/png', purpose: 'maskable' },
+          ],
+        },
+        workbox: {
+          globPatterns: ['**/*.{js,css,html,ico,png,svg,woff2}'],
+          navigateFallback: '/index.html',
+          navigateFallbackDenylist: [/^\/api/, /^\/firebase-messaging-sw\.js$/],
+        },
+      }),
+    ],
+    server: {
+      headers: {
+        'Cross-Origin-Opener-Policy': 'same-origin-allow-popups',
+      },
+    },
+    test: {
+      globals: true,
+      environment: 'jsdom',
+    },
+  }
 })
